requirements.txt
apify-client>=1.6.0
gspread>=6.0.0
pydantic>=2.0.0
PyYAML>=6.0
streamlit>=1.30.0
